pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "AI Coding Assistant CLI với Claude API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.84.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.1.0",
    "rich>=14.3.0",
]

[project.optional-dependencies]
  dev = [
      "ruff>=0.11.0",
  ]


[project.scripts]
code = "code_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/code_cli"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
